Move the FilesystemCommand to a trait.
